add another offsetlock in Chimera, will need to set the freq range of those ol channels

+++ Chimera/Source/OffsetLock/OlCore.h
@@ -43,7 +43,7 @@ public:
 
 	void resetConnection();
 private:
-	int tmp = 0;
+	unsigned long long tmp = 0;
 	std::array<std::string, size_t(OLGrid::total)> names;
 
 	//std::array <const double, 2> olResolution;
